feat: remove PWA install flow
