+++ docs/features/2026-02-08-pwa-update-toast.md
@@ -4,6 +4,9 @@ date: 2026-02-08
 status: implemented
 ---
 
+Deprecated: PWA support was removed on 2026-02-08. This update flow is no
+longer used. See `docs/features/2026-02-08-pwa-removal.md`.
+
 ## Summary
 
 Show a transient update bubble when a new service worker is available, then
