+++ docs/todo.md
@@ -27,9 +27,7 @@
 - [ ] Review raw HTML handling in markdown-it renderer and add sanitization if ACP output becomes untrusted (see `docs/features/2026-02-08-markdown-renderer.md`).
 - [ ] Validate SSE streaming output under proxy/load balancer conditions (see `docs/features/2026-02-08-sse-streaming.md`).
 - [ ] Verify SSE heartbeat events are emitted and ignored by the client (see `docs/features/2026-02-08-sse-streaming.md`).
-- [ ] Verify PWA install prompt, offline shell behavior, and Vite 7-compatible plugin version (see `docs/features/2026-02-08-pwa-support.md`).
-- [ ] Verify service worker registers only once with manual registration enabled (see `docs/features/2026-02-08-pwa-support.md`).
-- [ ] Verify PWA update toast appears and auto-refreshes on new service worker (see `docs/features/2026-02-08-pwa-update-toast.md`).
+- [ ] Verify push notifications still work with the minimal service worker after PWA removal (see `docs/features/2026-02-08-pwa-removal.md`).
 - [x] Verify SSE auto-reconnect and polling fallback after network interruption (see `docs/features/2026-02-08-sse-retry.md`).
 - [ ] Verify polling pauses while SSE is OPEN and resumes on disconnect (see `docs/features/2026-02-08-sse-retry.md`).
 - [ ] Verify SSE headers prevent proxy buffering and caching (see `docs/features/2026-02-08-sse-streaming.md`).
