+++ web/src/styles.css
@@ -157,26 +157,6 @@ h1 {
   color: #6c1b1b;
 }
 
-.sw-toast {
-  position: fixed;
-  right: 16px;
-  bottom: 16px;
-  z-index: 1000;
-  pointer-events: none;
-}
-
-.sw-toast-content {
-  display: inline-flex;
-  align-items: center;
-  gap: 10px;
-  padding: 8px 14px;
-  border-radius: 999px;
-  background: #111;
-  color: #fff;
-  font-size: 12px;
-  box-shadow: 0 10px 24px rgba(0, 0, 0, 0.2);
-}
-
 section {
   background: var(--panel);
   border: 1px solid var(--border);
