--- web/src/sw.js
@@ -1,43 +1,0 @@
-import { precacheAndRoute, cleanupOutdatedCaches } from "workbox-precaching";
-import { clientsClaim } from "workbox-core";
-import { registerRoute, NavigationRoute } from "workbox-routing";
-import { createHandlerBoundToURL } from "workbox-precaching";
-
-self.skipWaiting();
-clientsClaim();
-
-precacheAndRoute(self.__WB_MANIFEST ?? []);
-cleanupOutdatedCaches();
-
-const navigationHandler = createHandlerBoundToURL("/index.html");
-const navigationRoute = new NavigationRoute(navigationHandler, {
-  denylist: [/^\/api\//, /^\/sse\//, /\.[^/]+$/],
-});
-registerRoute(navigationRoute);
-
-self.addEventListener("push", (event) => {
-  let data = {};
-  try {
-    data = event.data ? event.data.json() : {};
-  } catch {
-    data = {};
-  }
-
-  const title = "AgentHub";
-  const body =
-    data.type === "agent_completed"
-      ? `Agent ${data.agent_id} completed`
-      : "You have a new notification";
-
-  event.waitUntil(
-    self.registration.showNotification(title, {
-      body,
-      data,
-    })
-  );
-});
-
-self.addEventListener("notificationclick", (event) => {
-  event.notification.close();
-  event.waitUntil(self.clients.openWindow("/"));
-});
